Fix MSVC warnings

+++ tests/vcall.cpp
@@ -153,8 +153,9 @@ Result vcall_impl(const char *domain, uint32_t n_inst, const Func &func,
 
     uint32_t se = jit_var_vcall(
         domain, self.index(), mask_combined.index(), n_inst, inst_id.data(),
-        indices_in.size(), indices_in.data(), indices_out_all.size(),
-        indices_out_all.data(), state.data(), indices_out.data());
+        (uint32_t) indices_in.size(), indices_in.data(),
+        (uint32_t) indices_out_all.size(), indices_out_all.data(), state.data(),
+        indices_out.data());
 
     jit_state.end_recording();
     jit_var_mark_side_effect(se);
